requirements.txt
clams-python==1.0.9
torch==2.*
torchvision
torchmetrics
opencv-python
tqdm
pyyaml
av==10.*